Folder restructure and renaming
