+++ src/Components/Pages/Team/data/data.js
@@ -1,17 +1,8 @@
 // TODO: Change the data once you get the details of your team members 
-export const data = [
-  {
-    name: "Saipranav Kotamreddy",
-    title: "Project Lead & BE Developer",
-    socials: {
-      ig: "https://instagram.com/aunsh___",
-      linkedIn: "https://instagram.com/in/aunsh",
-    },
-    profilePic: "https://i.postimg.cc/3NVyH6R8/SPKR.jpg",
-  },
+export const currentTeam = [
   {
     name: "Aunsh Bandivadekar",
-    title: "FE Developer & Architect",
+    title: "FE Developer & Co-Architect",
     socials: [
       "https://instagram.com/aunsh___",
       "https://linkedin.com/in/aunsh",
@@ -21,33 +12,57 @@ export const data = [
     quote: "Veni Vidi Vici",
   },
   {
-    name: "Nikita B. Emberi",
-    title: "Frontend Developer",
-    socials: [
-      "https://www.instagram.com/nikitaemberi/",
-      "https://www.linkedin.com/in/nikitaemberi/",
-      "https://github.com/NikitaEmberi/",
-    ],
-    profilePic: "https://i.postimg.cc/MKmv3BF2/nikita-resview.jpg",
-  },
-  {
-    name: "Madhumitha V.",
-    title: "Developer",
-    socials: [
-      "https://instagram.com/madhumitha1_?igshid=NzZlODBkYWE4Ng==",
-      "https://www.linkedin.com/in/vmadhuuu",
-      "https://github.com/vmadhuuu",
-    ],
-    profilePic: "https://i.postimg.cc/8zk9L8vQ/madhu.jpg",
-  },
-  {
-    name: "Sriharshini Dusi",
-    title: "Developer",
-    socials: [
-      "",
-      "",
-      " https://github.com/Sriharshinidusi",
-    ],
-    profilePic: "https://i.postimg.cc/9MPYx03x/harshini.jpg",
-  },
+    name: "Saipranav Kotamreddy",
+    title: "BE Developer & Co-Architect",
+    socials: {
+      ig: "https://instagram.com/aunsh___",
+      linkedIn: "https://instagram.com/in/aunsh",
+    },
+    profilePic: "https://i.postimg.cc/3NVyH6R8/SPKR.jpg",
+  }
 ];
+
+export const pastMembers = [
+  {},
+  {
+    name: "Dr. Mohammad Sadoghi",
+    title: "Principal & Guide",
+    socials: [
+      "",
+      "https://www.linkedin.com/in/mohammad-sadoghi/",
+      ""
+    ],
+    profilePic: "https://i.postimg.cc/3JCxDmRH/sadoghi.jpg",
+  },
+  {}
+  // {
+  //   name: "Nikita B. Emberi",
+  //   title: "Frontend Developer",
+  //   socials: [
+  //     "https://www.instagram.com/nikitaemberi/",
+  //     "https://www.linkedin.com/in/nikitaemberi/",
+  //     "https://github.com/NikitaEmberi/",
+  //   ],
+  //   profilePic: "https://i.postimg.cc/MKmv3BF2/nikita-resview.jpg",
+  // },
+  // {
+  //   name: "Madhumitha V.",
+  //   title: "Developer",
+  //   socials: [
+  //     "https://instagram.com/madhumitha1_?igshid=NzZlODBkYWE4Ng==",
+  //     "https://www.linkedin.com/in/vmadhuuu",
+  //     "https://github.com/vmadhuuu",
+  //   ],
+  //   profilePic: "https://i.postimg.cc/8zk9L8vQ/madhu.jpg",
+  // },
+  // {
+  //   name: "Sriharshini Dusi",
+  //   title: "Developer",
+  //   socials: [
+  //     "",
+  //     "",
+  //     " https://github.com/Sriharshinidusi",
+  //   ],
+  //   profilePic: "https://i.postimg.cc/9MPYx03x/harshini.jpg",
+  // },
+]
